Add SQLAlchemy Base and OfferRecord
